pyproject.toml
[project]
name = "ocr-checkbox"
version = "0.1.0"
description = "Checkbox detection and classification pipeline for PDF documents"
authors = [{name = "David Pepper", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core ML Framework
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "ultralytics>=8.0.0,<9.0.0",
    "timm>=0.6.0",
    
    # Computer Vision & PDF Processing
    "opencv-python>=4.5.0",
    "pdf2image>=1.16.0",
    "Pillow>=9.0.0",
    "pytesseract>=0.3.8",
    "PyMuPDF>=1.20.0",
    
    # Data Processing & Utilities
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "albumentations>=1.2.0",
    "scikit-learn>=1.1.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    
    # Configuration & Logging
    "pyyaml>=5.4.0",
    "tqdm>=4.60.0",
    # "wandb>=0.12.0",  # Optional - for experiment tracking
    
    # API & Deployment
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.5",
    
    # Additional utilities - can be added later if needed
    # "pycocotools>=2.0.4",  # May require additional setup on macOS
    # "supervision>=0.6.0",  # Optional visualization library
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "pytest>=6.2.0", 
    "black>=22.0.0",
    "flake8>=4.0.0",
    "pre-commit>=2.15.0",
]

# [project.scripts]
# train = "src.scripts.train:main"
# infer = "src.scripts.infer:main" 
# evaluate = "src.scripts.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
